+++ jobject-copy/src/main/java/de/alpharogroup/copy/object/CopyObjectExtensions.java
@@ -33,6 +33,8 @@ import java.util.Arrays;
 
 import org.apache.commons.beanutils.BeanUtils;
 
+import de.alpharogroup.lang.ClassType;
+import de.alpharogroup.lang.ObjectExtensions;
 import de.alpharogroup.reflection.ReflectionExtensions;
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
@@ -69,7 +71,7 @@ public final class CopyObjectExtensions
 	 *             is thrown if the class cannot be located
 	 */
 	public static <T> T copyObject(@NonNull T original)
-			throws IllegalAccessException, InstantiationException, ClassNotFoundException
+		throws IllegalAccessException, InstantiationException, ClassNotFoundException
 	{
 		T destination = ReflectionExtensions.newInstance(original);
 		return copyObject(original, destination);
@@ -104,8 +106,8 @@ public final class CopyObjectExtensions
 	 *             is thrown if the class cannot be located
 	 */
 	public static <ORIGINAL, DESTINATION> DESTINATION copyObject(@NonNull ORIGINAL original,
-																 @NonNull DESTINATION destination)
-			throws IllegalAccessException, InstantiationException, ClassNotFoundException
+		@NonNull DESTINATION destination)
+		throws IllegalAccessException, InstantiationException, ClassNotFoundException
 	{
 		for (Field field : original.getClass().getDeclaredFields())
 		{
@@ -148,13 +150,13 @@ public final class CopyObjectExtensions
 	 *             is thrown if the class cannot be located
 	 */
 	public static <ORIGINAL, DESTINATION> DESTINATION copyObject(@NonNull ORIGINAL original,
-																 @NonNull DESTINATION destination, String... ignoreFields)
-			throws IllegalAccessException, InstantiationException, ClassNotFoundException
+		@NonNull DESTINATION destination, String... ignoreFields)
+		throws IllegalAccessException, InstantiationException, ClassNotFoundException
 	{
 		for (Field field : original.getClass().getDeclaredFields())
 		{
 			if (Arrays.asList(ignoreFields).contains(field.getName())
-					|| copyField(field, original, destination))
+				|| copyField(field, original, destination))
 				continue;
 		}
 		return destination;
@@ -192,8 +194,8 @@ public final class CopyObjectExtensions
 	 */
 	@SuppressWarnings("unchecked")
 	public static <ORIGINAL, DESTINATION> boolean copyField(@NonNull Field field,
-															@NonNull ORIGINAL original, @NonNull DESTINATION destination)
-			throws IllegalAccessException, InstantiationException, ClassNotFoundException
+		@NonNull ORIGINAL original, @NonNull DESTINATION destination)
+		throws IllegalAccessException, InstantiationException, ClassNotFoundException
 	{
 		field.setAccessible(true);
 		Object value = field.get(original);
@@ -202,29 +204,31 @@ public final class CopyObjectExtensions
 			return true;
 		}
 		Class<?> fieldType = field.getType();
-		if (fieldType.isEnum())
+		ClassType classType = ObjectExtensions.getClassType(fieldType);
+		switch (classType)
 		{
-			Enum<?> enumValue = (Enum<?>)value;
-			String name = enumValue.name();
-			field.set(destination, Enum.valueOf(fieldType.asSubclass(Enum.class), name));
-		}
-		else if (fieldType.isPrimitive() || fieldType.equals(String.class)
-				|| fieldType.getSuperclass()!= null && fieldType.getSuperclass().equals(Number.class)
-				|| fieldType.equals(Boolean.class))
-		{
-			field.set(destination, value);
-		}
-		else
-		{
-			Object childObj = value;
-			if (childObj == original)
-			{
-				field.set(destination, destination);
-			}
-			else
-			{
-				field.set(destination, copyObject(value));
-			}
+			case ARRAY :
+				field.set(destination, ReflectionExtensions.copyArray(value));
+				break;
+			case ENUM :
+				Enum<?> enumValue = (Enum<?>)value;
+				String name = enumValue.name();
+				field.set(destination, Enum.valueOf(fieldType.asSubclass(Enum.class), name));
+				break;
+			case ANNOTATION :
+			case ANONYMOUS :
+			case COLLECTION :
+			case LOCAL :
+			case DEFAULT :
+			case MEMBER :
+			case SYNTHETIC :
+			case INTERFACE :
+			case PRIMITIVE :
+				field.set(destination, value);
+				break;
+			case MAP :
+				field.set(destination, value);
+				break;
 		}
 		return false;
 	}
@@ -252,8 +256,8 @@ public final class CopyObjectExtensions
 	 *             if the caller does not have access to the property accessor method
 	 */
 	public static <ORIGINAL, DESTINATION> DESTINATION copyPropertyWithReflection(
-			final ORIGINAL original, final DESTINATION destination, final String fieldName)
-			throws NoSuchFieldException, SecurityException, IllegalAccessException
+		final ORIGINAL original, final DESTINATION destination, final String fieldName)
+		throws NoSuchFieldException, SecurityException, IllegalAccessException
 	{
 		ReflectionExtensions.copyFieldValue(original, destination, fieldName);
 		return destination;
@@ -278,8 +282,8 @@ public final class CopyObjectExtensions
 	 *             if the property accessor method throws an exception
 	 */
 	public static <ORIGINAL, DESTINATION> DESTINATION copy(final ORIGINAL original,
-														   final DESTINATION destination)
-			throws IllegalAccessException, InvocationTargetException, IllegalArgumentException
+		final DESTINATION destination)
+		throws IllegalAccessException, InvocationTargetException, IllegalArgumentException
 	{
 		return copyProperties(original, destination);
 	}
@@ -303,7 +307,7 @@ public final class CopyObjectExtensions
 	 *             if the property accessor method throws an exception
 	 */
 	public static <ORIGINAL, DESTINATION> DESTINATION copyProperties(final ORIGINAL original,
-																	 final DESTINATION destination) throws IllegalAccessException, InvocationTargetException
+		final DESTINATION destination) throws IllegalAccessException, InvocationTargetException
 	{
 		BeanUtils.copyProperties(destination, original);
 		return destination;
@@ -337,7 +341,7 @@ public final class CopyObjectExtensions
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T> T copyProperties(final T original)
-			throws IllegalAccessException, InvocationTargetException, InstantiationException
+		throws IllegalAccessException, InvocationTargetException, InstantiationException
 	{
 		Object destination = original.getClass().newInstance();
 		BeanUtils.copyProperties(destination, original);
@@ -361,16 +365,16 @@ public final class CopyObjectExtensions
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T extends Serializable> T copySerializedObject(final T orig)
-			throws IOException, ClassNotFoundException
+		throws IOException, ClassNotFoundException
 	{
 		try (ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-			 ObjectOutputStream objectOutputStream = new ObjectOutputStream(byteArrayOutputStream))
+			ObjectOutputStream objectOutputStream = new ObjectOutputStream(byteArrayOutputStream))
 		{
 			objectOutputStream.writeObject(orig);
 			objectOutputStream.flush();
 			objectOutputStream.close();
 			final ByteArrayInputStream bis = new ByteArrayInputStream(
-					byteArrayOutputStream.toByteArray());
+				byteArrayOutputStream.toByteArray());
 			final ObjectInputStream ois = new ObjectInputStream(bis);
 			T object = (T)ois.readObject();
 			byteArrayOutputStream.close();